refact: 파일 구조 세분화
hooks 폴더 생성 및 파일 구조 세분화
공부 성향 조사 로직 hooks 폴더로 관리
